Update renamed/moved imports
