+++ styles/styleTree/workspace.ts
@@ -1,5 +1,5 @@
+import Theme from "../themes/theme";
 import { backgroundColor, border, iconColor, text } from "./components";
-import Theme from "./theme";
 
 export default function workspace(theme: Theme) {
   const signInPrompt = {
